+++ Runtime/Scripts/INotifyDataSourceChanged.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace de.JochenHeckl.Unity.DataBinding
+{
+    public interface INotifyDataSourceChanged
+	{
+		event Action DataSourceChanged;
+	}
+}