+++ Samples~/ViewsAndDataSources/Scripts/ElementDataSource.cs
@@ -1,0 +1,7 @@
+namespace de.JochenHeckl.Unity.DataBinding.Example
+{
+    public class ElementDataSource : DataSourceBase<ElementDataSource>
+    {
+        public string ElementValue { get; set; }
+    }
+}